backend_python/requirements.txt
# Zaragoza Histórica - Backend Python Dependencies
# FastAPI y servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# Base de datos
psycopg2-binary>=2.9.9

# Utilidades
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Procesamiento de imágenes (validación y thumbnails)
Pillow>=10.4.0

# Rate limiting
slowapi>=0.1.9

# CORS y seguridad
python-multipart>=0.0.6

# Autenticación JWT
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Importación GML del Catastro (sólo necesario para scripts/import_buildings.py)
lxml>=5.0.0
pyproj>=3.6.0
